# build/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "mongoose.Types.ObjectId": {
        "type": "string"
      },
      "AccountState": {
        "enum": ["in-used", "forbidden"],
        "type": "string"
      },
      "UserDTO": {
        "properties": {
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录创建时间"
          },
          "modifyTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录修改时间"
          },
          "deleteTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "记录删除时间"
          },
          "id": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "abc123!@#$"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "kbmjj123@gmail.com"
          },
          "role": {
            "type": "string",
            "enum": ["user", "admin"],
            "description": "用户角色，枚举类型，user代表普通用户，admin代表管理员"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true,
            "description": "刷新的token"
          },
          "accessToken": {
            "type": "string",
            "nullable": true,
            "description": "资源访问的token"
          },
          "nickName": {
            "type": "string",
            "nullable": true,
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "nullable": true,
            "description": "用户头像"
          },
          "address": {
            "allOf": [
              {
                "$ref": "#/components/schemas/mongoose.Types.ObjectId"
              }
            ],
            "nullable": true,
            "description": "用户地址id"
          },
          "loginTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "登录时间"
          },
          "logoutTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "退出时间"
          },
          "account": {
            "type": "string",
            "nullable": true,
            "description": "用户账号"
          },
          "state": {
            "$ref": "#/components/schemas/AccountState",
            "description": "用户状态"
          }
        },
        "required": ["id", "password", "email", "role"],
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_UserDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/UserDTO"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "Record_string.any_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
      "LanguageItemType": {
        "properties": {
          "languageMap": {
            "$ref": "#/components/schemas/Record_string.any_"
          },
          "language": {
            "type": "string"
          }
        },
        "required": ["languageMap", "language"],
        "type": "object"
      },
...
      "BaseObjectEntity_string_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_unknown_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {}
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "CateDTO": {
        "properties": {
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录创建时间"
          },
          "modifyTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录修改时间"
          },
          "deleteTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "记录删除时间"
          },
          "languageList": {
            "items": {
              "$ref": "#/components/schemas/LanguageItemType"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "parentId": {
            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
          },
          "level": {
            "type": "number",
            "format": "double"
          },
          "paramsList": {
            "items": {
              "properties": {
                "values": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "key": {
                  "type": "string"
                }
              },
              "required": ["values", "key"],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": ["languageList", "title"],
        "type": "object",
        "additionalProperties": false
      },
...
      "BaseObjectEntity_CateDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/CateDTO"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Pick_CateDTO.title-or-level-or-parentId-or-paramsList_": {
        "properties": {
          "title": {
            "type": "string"
          },
          "level": {
            "type": "number",
            "format": "double"
          },
          "parentId": {
            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
          },
          "paramsList": {
            "items": {
              "properties": {
                "values": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "key": {
                  "type": "string"
                }
              },
              "required": ["values", "key"],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": ["title"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Partial_Pick_CateDTO.id__": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "type": "object",
        "description": "Make all properties in T optional"
      },
      "EditCateDTO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_CateDTO.title-or-level-or-parentId-or-paramsList_"
          },
          {
            "$ref": "#/components/schemas/Partial_Pick_CateDTO.id__"
          }
        ],
        "description": "新增/编辑分类的参数"
      },
      "BaseObjectEntity_CateDTO-or-null_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CateDTO"
              }
            ],
            "nullable": true
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "BrandDTO": {
        "properties": {
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录创建时间"
          },
          "modifyTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录修改时间"
          },
          "deleteTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "记录删除时间"
          },
          "languageList": {
            "items": {
              "$ref": "#/components/schemas/LanguageItemType"
            },
            "type": "array"
          },
          "id": {
            "type": "string",
            "description": "id"
          },
          "name": {
            "type": "string",
            "description": "品牌名称"
          },
          "icon": {
            "type": "string",
            "description": "品牌图标"
          }
        },
        "required": ["languageList", "id", "name", "icon"],
        "type": "object",
        "additionalProperties": false
      },
      "PageListType_BrandDTO_": {
        "properties": {
          "pages": {
            "type": "number",
            "format": "double"
          },
          "pageIndex": {
            "type": "number",
            "format": "double"
          },
          "pageSize": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "list": {
            "items": {
              "$ref": "#/components/schemas/BrandDTO"
            },
            "type": "array"
          }
        },
        "required": ["pages", "pageIndex", "pageSize", "total", "list"],
        "type": "object"
      },
      "BasePageListEntity_BrandDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageListType_BrandDTO_"
          }
        },
        "required": ["data"],
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_BrandDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/BrandDTO"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Pick_BrandDTO.name-or-icon_": {
        "properties": {
          "name": {
            "type": "string",
            "description": "品牌名称"
          },
          "icon": {
            "type": "string",
            "description": "品牌图标"
          }
        },
        "required": ["name", "icon"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Partial_Pick_BrandDTO.languageList__": {
        "properties": {
          "languageList": {
            "items": {
              "$ref": "#/components/schemas/LanguageItemType"
            },
            "type": "array"
          }
        },
        "type": "object",
        "description": "Make all properties in T optional"
      },
      "EditBrandDTO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Pick_BrandDTO.name-or-icon_"
          },
          {
            "$ref": "#/components/schemas/Partial_Pick_BrandDTO.languageList__"
          }
        ],
        "description": "发布品牌所需的参数定义--多语言支持"
      },
      "BaseObjectEntity_BrandDTO-or-null_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BrandDTO"
              }
            ],
            "nullable": true
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "PageListType_UserDTO_": {
        "properties": {
          "pages": {
            "type": "number",
            "format": "double"
          },
          "pageIndex": {
            "type": "number",
            "format": "double"
          },
          "pageSize": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "list": {
            "items": {
              "$ref": "#/components/schemas/UserDTO"
            },
            "type": "array"
          }
        },
        "required": ["pages", "pageIndex", "pageSize", "total", "list"],
        "type": "object"
      },
      "BasePageListEntity_UserDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageListType_UserDTO_"
          }
        },
        "required": ["data"],
        "type": "object",
        "additionalProperties": false
      },
...
      "Omit_UserDTO.password-or-refreshToken-or-accessToken-or-logoutTime-or-isPasswordMatched_": {
        "$ref": "#/components/schemas/Pick_UserDTO.Exclude_keyofUserDTO.password-or-refreshToken-or-accessToken-or-logoutTime-or-isPasswordMatched__",
        "description": "Construct a type with the properties of T except for those in type K."
      },
      "UserWithoutToken": {
        "$ref": "#/components/schemas/Omit_UserDTO.password-or-refreshToken-or-accessToken-or-logoutTime-or-isPasswordMatched_",
        "description": "不带token的用户信息"
      },
      "BaseObjectEntity_UserWithoutToken_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/UserWithoutToken"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Pick_UserDTO.state_": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/AccountState",
            "description": "用户状态"
          }
        },
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "UserToggleEnabledParams": {
        "$ref": "#/components/schemas/Pick_UserDTO.state_",
        "description": "变更用户状态的参数"
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "KBMall",
    "version": "1.0.0",
    "description": "KB商城的买家端、卖家端的接口服务",
    "license": {
      "name": "ISC"
    },
    "contact": {
      "email": "kbmjj123@gmail.com",
      "name": "郑耿林"
    }
  },
  "paths": {
...
    "/cate": {
      "put": {
        "operationId": "AddACate",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_CateDTO_"
                }
              }
            }
          }
        },
        "tags": ["分类模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCateDTO"
              }
            }
          }
        }
      }
    },
    "/cate/{id}": {
      "post": {
        "operationId": "EditACate",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_CateDTO-or-null_"
                }
              }
            }
          }
        },
        "description": "编辑一分类信息",
        "tags": ["分类模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CateDTO"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "RemoveACate",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_unknown_"
                }
              }
            }
          }
        },
        "description": "删除一分类信息",
        "tags": ["分类模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/brand/list": {
      "get": {
        "operationId": "GetBrandList",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasePageListEntity_BrandDTO_"
                }
              }
            }
          }
        },
        "description": "获取品牌列表",
        "tags": ["品牌模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "pageIndex",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/brand": {
      "put": {
        "operationId": "AddABrand",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_BrandDTO_"
                }
              }
            }
          }
        },
        "description": "新增一品牌",
        "tags": ["品牌模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditBrandDTO"
              }
            }
          }
        }
      }
    },
    "/brand/{id}": {
      "get": {
        "operationId": "GetABrand",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_BrandDTO_"
                }
              }
            }
          }
        },
        "description": "获取一品牌信息",
        "tags": ["品牌模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "EditABrand",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_BrandDTO-or-null_"
                }
              }
            }
          }
        },
        "description": "编辑一品牌",
        "tags": ["品牌模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditBrandDTO"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "RemoveABrand",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_string_"
                }
              }
            }
          }
        },
        "description": "删除品牌",
        "tags": ["品牌模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/account/list": {
      "get": {
        "operationId": "GetUserList",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasePageListEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "获取用户列表",
        "tags": ["账号模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "pageIndex",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/account/info/{id}": {
      "get": {
        "operationId": "GetAUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserWithoutToken_"
                }
              }
            }
          }
        },
        "description": "根据id获取用户信息",
        "tags": ["账号模块"],
        "security": [],
        "parameters": [
          {
            "description": "用户id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/account/{id}/toggleAccountState": {
      "post": {
        "operationId": "ToggleAccountState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_UserDTO_"
                }
              }
            }
          }
        },
        "description": "启用/禁用一账号",
        "tags": ["账号模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserToggleEnabledParams"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}